Add tests for migrating old location db to new
