+++ elodie/geolocation.py
@@ -127,7 +127,7 @@ def place_name(lat, lon):
     cached_place_name = db.get_location_name(lat, lon, 3000)
     # We check that it's a dict to coerce an upgrade of the location
     #  db from a string location to a dictionary. See gh-160.
-    if(cached_place_name):
+    if(isinstance(cached_place_name, dict)):
         return cached_place_name
 
     lookup_place_name = {}
