+++ elodie/tests/elodie_test.py
@@ -14,7 +14,6 @@ sys.path.insert(0, os.path.abspath(os.path.dirname(os.path.dirname(os.path.dirna
 import helper
 elodie = load_source('elodie', os.path.abspath('{}/../../elodie.py'.format(os.path.dirname(os.path.realpath(__file__)))))
 
-from elodie import constants
 from elodie.localstorage import Db
 from elodie.media.audio import Audio
 from elodie.media.photo import Photo
@@ -31,9 +30,9 @@ def test_import_file_text():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -47,9 +46,9 @@ def test_import_file_audio():
     origin = '%s/audio.m4a' % folder
     shutil.copyfile(helper.get_file('audio.m4a'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -63,9 +62,9 @@ def test_import_file_photo():
     origin = '%s/plain.jpg' % folder
     shutil.copyfile(helper.get_file('plain.jpg'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -79,9 +78,9 @@ def test_import_file_video():
     origin = '%s/video.mov' % folder
     shutil.copyfile(helper.get_file('video.mov'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -95,10 +94,10 @@ def test_import_file_allow_duplicate_false():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path1 = elodie.import_file(origin, folder_destination, False, False, False)
     dest_path2 = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -113,10 +112,10 @@ def test_import_file_allow_duplicate_true():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path1 = elodie.import_file(origin, folder_destination, False, False, True)
     dest_path2 = elodie.import_file(origin, folder_destination, False, False, True)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
     shutil.rmtree(folder_destination)
@@ -133,9 +132,9 @@ def test_import_destination_in_source():
     origin = '%s/video.mov' % folder
     shutil.copyfile(helper.get_file('video.mov'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     dest_path = elodie.import_file(origin, folder_destination, False, False, False)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
 
@@ -151,9 +150,9 @@ def test_update_location_on_audio():
     audio = Audio(origin)
     metadata = audio.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_location(audio, origin, 'Sunnyvale, CA')
-    restore_dbs()
+    helper.restore_dbs()
 
     audio_processed = Audio(origin)
     metadata_processed = audio_processed.get_metadata()
@@ -176,9 +175,9 @@ def test_update_location_on_photo():
     photo = Photo(origin)
     metadata = photo.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_location(photo, origin, 'Sunnyvale, CA')
-    restore_dbs()
+    helper.restore_dbs()
 
     photo_processed = Photo(origin)
     metadata_processed = photo_processed.get_metadata()
@@ -201,9 +200,9 @@ def test_update_location_on_text():
     text = Text(origin)
     metadata = text.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_location(text, origin, 'Sunnyvale, CA')
-    restore_dbs()
+    helper.restore_dbs()
 
     text_processed = Text(origin)
     metadata_processed = text_processed.get_metadata()
@@ -226,9 +225,9 @@ def test_update_location_on_video():
     video = Video(origin)
     metadata = video.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_location(video, origin, 'Sunnyvale, CA')
-    restore_dbs()
+    helper.restore_dbs()
 
     video_processed = Video(origin)
     metadata_processed = video_processed.get_metadata()
@@ -251,9 +250,9 @@ def test_update_time_on_audio():
     audio = Audio(origin)
     metadata = audio.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_time(audio, origin, '2000-01-01 12:00:00')
-    restore_dbs()
+    helper.restore_dbs()
 
     audio_processed = Audio(origin)
     metadata_processed = audio_processed.get_metadata()
@@ -275,9 +274,9 @@ def test_update_time_on_photo():
     photo = Photo(origin)
     metadata = photo.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_time(photo, origin, '2000-01-01 12:00:00')
-    restore_dbs()
+    helper.restore_dbs()
 
     photo_processed = Photo(origin)
     metadata_processed = photo_processed.get_metadata()
@@ -299,9 +298,9 @@ def test_update_time_on_text():
     text = Text(origin)
     metadata = text.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_time(text, origin, '2000-01-01 12:00:00')
-    restore_dbs()
+    helper.restore_dbs()
 
     text_processed = Text(origin)
     metadata_processed = text_processed.get_metadata()
@@ -323,9 +322,9 @@ def test_update_time_on_video():
     video = Video(origin)
     metadata = video.get_metadata()
 
-    reset_dbs()
+    helper.reset_dbs()
     status = elodie.update_time(video, origin, '2000-01-01 12:00:00')
-    restore_dbs()
+    helper.restore_dbs()
 
     video_processed = Video(origin)
     metadata_processed = video_processed.get_metadata()
@@ -348,11 +347,11 @@ def test_regenerate_valid_source():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     runner = CliRunner()
     result = runner.invoke(elodie._generate_db, ['--source', folder])
     db = Db()
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
 
@@ -367,11 +366,11 @@ def test_regenerate_valid_source_with_invalid_files():
     origin_invalid = '%s/invalid.invalid' % folder
     shutil.copyfile(helper.get_file('invalid.invalid'), origin_invalid)
 
-    reset_dbs()
+    helper.reset_dbs()
     runner = CliRunner()
     result = runner.invoke(elodie._generate_db, ['--source', folder])
     db = Db()
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
 
@@ -385,11 +384,11 @@ def test_verify_ok():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     runner = CliRunner()
     runner.invoke(elodie._generate_db, ['--source', folder])
     result = runner.invoke(elodie._verify)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
 
@@ -402,33 +401,15 @@ def test_verify_error():
     origin = '%s/valid.txt' % folder
     shutil.copyfile(helper.get_file('valid.txt'), origin)
 
-    reset_dbs()
+    helper.reset_dbs()
     runner = CliRunner()
     runner.invoke(elodie._generate_db, ['--source', folder])
     with open(origin, 'w') as f:
         f.write('changed text')
     result = runner.invoke(elodie._verify)
-    restore_dbs()
+    helper.restore_dbs()
 
     shutil.rmtree(folder)
 
     assert origin in result.output, result.output
     assert 'Error           1' in result.output, result.output
-
-def reset_dbs():
-    hash_db = constants.hash_db
-    if os.path.isfile(hash_db):
-        os.rename(hash_db, '{}-test'.format(hash_db))
-
-    location_db = constants.location_db
-    if os.path.isfile(location_db):
-        os.rename(location_db, '{}-test'.format(location_db))
-
-def restore_dbs():
-    hash_db = '{}-test'.format(constants.hash_db)
-    if os.path.isfile(hash_db):
-        os.rename(hash_db, hash_db.replace('-test', ''))
-
-    location_db = '{}-test'.format(constants.location_db)
-    if os.path.isfile(location_db):
-        os.rename(location_db, location_db.replace('-test', ''))
